Modified getBatches method to return a plain array instead of a collection
When doing an email retry, the sendEmail function is first called with the output of createBatches (an array), then upon retrying it is called with the output of the getBatches method (a collection).
This caused an error on the retry, because a collection doesn't have the .indexOf() method.
This change aligns these two methods so they return the same type of object.